requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# API clients
anthropic>=0.31.0
httpx>=0.24.0
tenacity>=8.2.0

# Data processing
openpyxl>=3.1.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Report generation
markdown>=3.4.0
jinja2>=3.1.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.0.0
ruff>=0.0.280
mypy>=1.4.0
python-dotenv>=1.0.0

# Type stubs
pandas-stubs>=2.0.0
types-pyyaml>=6.0.0